==> 1_data_prep.Rmd <==
---
title: "Perm Mortality Data Preparation"
author: "Mike McPhee Anderson, FSA"
date: "2024-12-14"
output: 
  html_document:
    css: ["squareerror_markdown.css"]
    includes:
      in_header: "squareerror_header.html"
---

```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = TRUE)

library(RPostgreSQL)
library(tidyverse)

```

## Prepare ILEC data extract

This notebook extracts the "Perm" (whole life) subset of data from the much larger
[ILEC data](https://www.soa.org/resources/research-reports/2021/2017-mortality-experience/) data.  I've imported the data into a Postgres database to make it more managable.

* An alternative is to download the CSV from the ILEC page, and import the CSV into DuckDB. The only thing that would change is the `DBI::dbConnect` statement, as the rest of the R code is independent of the underlying database.

```{r warning=F}

if (!exists("conn")) {
  
  # the database is running on a local (non-exposed) virtual network via Docker,
  # please excuse the plaintext password stored in GitHub 
  conn <- DBI::dbConnect(dbDriver("PostgreSQL"), 
                         dbname = "postgres", 
                         host="postgis", 
                         port=5432, 
                         user="postgres", 
                         password="postgres")
  
  # remove the next line if you are not using postgres, or if it 
  # does not match your database structure
  DBI::dbExecute(conn, "SET SEARCH_PATH='soa'")
}

tbl_ilec <- tbl(conn, "ilec_2009_2018") %>%
  mutate(
    gender = upper(trim(gender)),
    smoker_status =  upper(trim(smoker_status)),
    insurance_plan = upper(trim(insurance_plan)),
    age_basis = upper(trim(age_basis)),
    face_amount_band = upper(trim(face_amount_band)),
    number_of_preferred_classes = upper(trim(number_of_preferred_classes)),
    preferred_class = upper(trim(preferred_class)),
    table_version = "2008 VBT",
    table_name = case_when(
      number_of_preferred_classes == "NA" ~ "LIMITED UNDERWRITING",
      TRUE ~ "PRIMARY"
    )
  ) %>%
  filter(
    number_of_preferred_classes != "4",
    insurance_plan == "PERM",
    smoker_status != "UNKNOWN")

head(tbl_ilec) 

```

## Append ultimate mortality

Most life insurance has a *select* period following underwriting. Mortality is 
better during this period, since underwriting has filtered out the worst risks.
The risk filtering effects diminish as the time since underwriting increases, and 
when it is fully diminished, mortality from the *ultimate* table is used.

This exercise uses an ultimate table mortality as an "offset" term, which relieves
the model from having to estimate age and sex specific effects, and instead allows
it to focus more on insurance attributes of interest (underwriting class, time since underwriting, etc.).

```{r}

tbl_vbt <- tbl(conn, "vbt_tables") %>%
  filter(table_version == "2008 VBT",
         select_period == F,
         issue_age == 0,
         table_name %in% c("PRIMARY", "LIMITED UNDERWRITING"))

tbl_ilec.qx <- tbl_ilec %>%
  left_join(tbl_vbt %>% select(-issue_age), by=c(
    "attained_age" = "attained_age",
    "gender"="gender",
    "smoker_status"="smoker",
    "table_version"="table_version", 
    "table_name"="table_name", 
    "age_basis"="alb_anb"))

```

### Sanity checks for the data extract

```{r warning=F}

tbl_ilec %>%
  summarise(
    cnt_claims = sum(number_of_deaths)
  ) %>%
  knitr::kable()

```


```{r warning=F}

tbl_ilec.qx %>%
  filter(!is.na(qx)) %>%
  summarise(
    cnt_claims = sum(number_of_deaths)
  ) %>%
  knitr::kable()

```

## Create Extract

```{r}

tbl_ilec.qx %>%
  select(observation_year, preferred_indicator, gender, issue_age, duration,
         attained_age, age_basis, face_amount_band, issue_year, smoker_status,
         number_of_preferred_classes, preferred_class,
         number_of_deaths, death_claim_amount, 
         policies_exposed, amount_exposed, qx) %>%
  collect() %>%
  write_rds("/mnt/data/ilec/perm_extract.rds")

```

```{r}

DBI::dbDisconnect(conn)
rm(conn)

```
