==> 2_perm_mortality_eda.Rmd <==
---
title: "Perm Mortality EDA"
author: "Mike McPhee Anderson, FSA"
date: "2024-12-14"
output: 
  html_document:
    css: ["squareerror_markdown.css"]
    includes:
      in_header: "squareerror_header.html"
---

```{r setup, include=FALSE}

library(glmnet)
library(rpart)
library(rpart.plot)
library(splines)
library(tidyverse)

```

## Load Study Data

Load the data created in [part 1](https://github.com/mma-fsa/perm_mortality/blob/main/1_data_prep.Rmd).

 * In actuarial nomenclature, we call insurance claims data *experience study* data or just *study* data for short.

```{r}

df_study_data <- read_rds("/mnt/data/ilec/perm_extract.rds")

```

## Append additional columns to study data

* Setup the face amount as an ordered factor

  * This is useful for some models, like decision trees, which can make use of ordinal factors
 
  * Otherwise, this can easily be cast as an integer datatype for models (like GLMs), when a linear predictor is desired.
 
* The face bands are also a bit too granular, especially at high and low amounts, so create a higher-level grouping. 


```{r}

face_amount_order <- c(
  "1-9999",
  "10000-24999",
  "25000-49999",
  "50000-99999",
  "100000-249999",
  "250000-499999",
  "500000-999999",
  "1000000-2499999",
  "2500000-4999999",
  "5000000-9999999",
  "10000000+"
)

df_study_data <- df_study_data %>%
  mutate(
    face_amount_band = ordered(face_amount_band, levels=!!face_amount_order),
    face_amount_band_int = as.integer(face_amount_band),
    face_amount_band_grp = forcats::fct_collapse(face_amount_band,
      `<$25K` = c("1-9999", "10000-24999"),
      `$25K-$99K` = c("25000-49999", "50000-99999"),
      `$100K-$499K` = c("100000-249999", "250000-499999"),
      `$500K-$999K` = c("500000-999999"),
      `$1M-$2.499M` = c("1000000-2499999"),
      `$2.5M-4.99M` = c("2500000-4999999"),
      `$5M+` = c("5000000-9999999", "10000000+")
    )
  )

```


## EDA

### Check Exposure by UW Class

In actuarial (and demographer) nomenclature, an *exposure* is a single year for a given life.
For example, if you had 50 lives at a given age, and one of them died, the mortality rate ($q_x$) is simply the deaths divided by exposures ($1 \div 50 = 0.02 = q_x$).

We need to understand how much exposure (data) we have for each underwriting class. An underwriting class is based on the insurance company's assessment of a policyholders health, at the time the policy is issued. This will inform the structure of any model that we build, since we may need to group underwriting classes when there is low amounts of exposure.

 * Over time, the number of underwriting classes has increased, leading to more granular separation of risks. Unfortunately, this leads to problems in predictive modeling since the "best" underwriting class today is much better than it was decades ago, because it's harder to get the best class when there are 4 possible underwriting classes vs. 2. 
 
   * With more recent data, there will be 3-4 classes for non-smokers, and 2 for smokers.
  
   * For older data, there will be 2 classes for non-smokers, and 1 for smokers.
  
 * There are less stringent forms of underwriting, called simplified issue, that issue only non-smoker vs. smoker underwriting classes.  These policies are typically lower face amount.  There appears to be a sizable cohort of simplified issue business in this data.

```{r}

df_study_data %>% 
  filter(!is.na(qx)) %>%
  group_by(number_of_preferred_classes, preferred_class, smoker_status, gender) %>%
  summarise(
    exposure = sum(policies_exposed),
    deaths = sum(number_of_deaths),
    expected = sum(policies_exposed * qx),
    .groups="drop"
  ) %>%
  mutate(
    `A/E` = deaths / expected
  ) %>%
  arrange(desc(exposure)) %>%
  knitr::kable(digits = 2)

```

### Initial A/E Plots

Actuaries typically examine the actual deaths vs. expected deaths (A/E).

 * This is somewhat related to the Poisson likelihood used for predictive modeling, where the model's objective is  $\text{argmax}_{\theta}\ L(\text{Actual} | \theta)$, where $\theta$ creates an expected/predicted amount. In other words, like the A/E ratio, the Poisson likelihood function takes actual and expected as input.
 
* The below plots show us where the existing basis is doing well, v.s. where the GLM will need to heavily adjust things. Generally speaking, younger and older ages are the closest to to 100% A/E (ideal); this is good because this is typically where data volumes are the lowest.


```{r fig.width=10, warning=F}

plot_data <- df_study_data %>%
  mutate(
    uw_class_full = paste0("n=", number_of_preferred_classes, " cls=", preferred_class, " smk=", smoker_status)
  ) %>%
  filter(!is.na(qx)) %>%
  group_by(attained_age, uw_class_full, gender) %>%
  summarise(
    exposure = sum(policies_exposed),
    deaths = sum(number_of_deaths),
    expected = sum(policies_exposed * qx),
    .groups="drop"
  ) %>%
  mutate(
    ae = pmin(2, deaths / expected)
  ) %>%
  arrange(ae)

ggplot(plot_data, aes(x=attained_age, y=ae, color=gender, group=gender)) +
  facet_wrap(~ uw_class_full, scales="free_y") +
  scale_y_continuous(trans=scales:::log_trans(), labels = scales::label_percent()) +
  geom_smooth(se = F, span=0.5) +
  geom_hline(yintercept = 1) + 
  theme_bw() +
  ylab("A/E") +
  ggtitle("Actual / Expected - by UW Class", "Expected basis is 2008 VBT ultimate table")

```

## Three nonsmoker classes, by duration

* The data only supports a 10-15 year selection periods for all the classes

```{r fig.width=8, fig.height=3}

plot_data <- df_study_data %>%
  filter(number_of_preferred_classes == "3") %>%
  group_by(gender, preferred_class, duration) %>%
  summarise(
    exposure = sum(policies_exposed),
    deaths = sum(number_of_deaths),
    expected = sum(policies_exposed * qx),
    .groups = "drop"
  )

ggplot(plot_data, aes(x=duration, y=deaths, fill=gender, group=gender)) +
  facet_wrap(~ preferred_class, nrow = 1) +
  geom_bar(stat="identity", position = "dodge") +
  theme_bw() +
  ggtitle("Non-smoker, 3 underwriting classes")

```

## Two nonsmoker classes, by duration

* These would support up to a 25 year selection period

```{r fig.width=8, fig.height=3}

plot_data <- df_study_data %>%
  filter(number_of_preferred_classes == "2", smoker_status=="NONSMOKER") %>%
  group_by(gender, preferred_class, duration) %>%
  summarise(
    exposure = sum(policies_exposed),
    deaths = sum(number_of_deaths),
    expected = sum(policies_exposed * qx),
    .groups="drop"
  )

ggplot(plot_data, aes(x=duration, y=deaths, fill=gender, group=gender)) +
  facet_wrap(~ preferred_class, nrow=1) +
  geom_bar(stat="identity", position = "dodge") +
  theme_bw() +
  ggtitle("Non-smoker, 2 underwriting classes")

```

## Two smoker classes, by duration

* Pretty low data volumes, but a 25 year select period isn't out of the question.

```{r fig.width=8, fig.height=3}

plot_data <- df_study_data %>%
  filter(number_of_preferred_classes == "2", smoker_status=="SMOKER") %>%
  group_by(gender, preferred_class, duration) %>%
  summarise(
    exposure = sum(policies_exposed),
    deaths = sum(number_of_deaths),
    expected = sum(policies_exposed * qx),
    .groups="drop"
  )

ggplot(plot_data, aes(x=duration, y=deaths, fill=gender, group=gender)) +
  facet_wrap(~ preferred_class, nrow=1) +
  geom_bar(stat="identity", position = "dodge") +
  theme_bw() +
  ggtitle("Smoker, 2 underwriting classes")

```


## Smoker Distinct (one class)

* Bulk of the data is late duration, so it is likely very old fully underwriting business.

* More recently issued policies are probably simplified type policies

```{r fig.width=8, fig.height=3}

plot_data <- df_study_data %>%
  filter(number_of_preferred_classes == "NA") %>%
  group_by(gender, smoker_status, duration) %>%
  summarise(
    exposure = sum(policies_exposed),
    deaths = sum(number_of_deaths),
    expected = sum(policies_exposed * qx),
    .groups="drop"
  )

ggplot(plot_data, aes(x=duration, y=deaths, fill=gender, group=gender)) +
  facet_wrap(~ smoker_status, nrow=1) +
  geom_bar(stat="identity", position = "dodge") +
  theme_bw() +
  ggtitle("Smoker distinct")


```

### Face Amount vs. Issue Year (smoker distinct, NS only)

Check for final-expense / simplified issue type policies

* Based on the cliff at the higher face amounts, looks like mostly simplfied issue for 2005 and after

* Weird spike around 2015

```{r fig.width=10}

plot_data <- df_study_data %>%
  filter(number_of_preferred_classes == "NA",
         smoker_status == "NONSMOKER") %>%
  group_by(face_amount_band_grp, issue_year) %>%
  summarise(
    exposure = sum(policies_exposed),
    deaths = sum(number_of_deaths),
    expected = sum(policies_exposed * qx),
    .groups="drop"
  )

ggplot(plot_data, aes(x=issue_year, y=exposure)) +
  facet_wrap(~ face_amount_band_grp, scales="free_y") +
  geom_bar(stat="identity") +
  geom_vline(xintercept = 2004, color="red") +
  geom_vline(xintercept = 2015, color="blue") +
  ggtitle("Simplified issue vs. Fully Underwritten", 
          "Try and determine an issue-year cutoff point")

```

### Face Amount vs. Issue Year (smoker distinct, smokers only)

* Pattern still present at higher face amounts (post 2005 cliff), but less evident.

```{r fig.width=10}

plot_data <- df_study_data %>%
  filter(number_of_preferred_classes == "NA",
         smoker_status == "SMOKER") %>%
  group_by(face_amount_band_grp, issue_year) %>%
  summarise(
    exposure = sum(policies_exposed),
    deaths = sum(number_of_deaths),
    expected = sum(policies_exposed * qx),
    .groups="drop"
  )

ggplot(plot_data, aes(x=issue_year, y=exposure)) +
  facet_wrap(~ face_amount_band_grp, scales="free_y") +
  geom_bar(stat="identity") +
  geom_vline(xintercept = 2004, color="red")

```

### Face amount vs. Issue Year (2 class NS)

* Maybe some 2-class SI in \$25K-\$99K

```{r fig.width=10}

plot_data <- df_study_data %>%
  filter(number_of_preferred_classes == "2",
         smoker_status == "NONSMOKER") %>%
  group_by(face_amount_band_grp, issue_year) %>%
  summarise(
    exposure = sum(policies_exposed),
    deaths = sum(number_of_deaths),
    expected = sum(policies_exposed * qx),
    .groups="drop"
  )

ggplot(plot_data, aes(x=issue_year, y=exposure)) +
  facet_wrap(~ face_amount_band_grp, scales="free_y") +
  geom_bar(stat="identity") 

```

## Create Data Splits

The exploratory data analysis points to a couple major segments of business. With more predictors to work with, we could perhaps model these together in a single model.  But, since we have limited information about the underlying business, it's easier to model these separately. 

 1. **Perm Historical (PH)**:  Policies prior to 2005 are likely to have little remaining effects from underwriting selection.  Also, this data is unlikely to have preferred classes, and is mostly just smoker vs. non-smoker for underwriting classes.  The majority of it will be smaller face amounts. 

 2. **Simplified Issue:** Policies issued after 2005 without any preferred classes are likely simplified
   issue.  These policies are less than $100K face amount, and are likely to have less stringent underwriting.
 
 3. **Fully Underwritten (recent):**  Post-2005 policies are likely to have at least one preferred class, and I recall that blood testing gained traction in the late 1990s.  These policies will also have some underwriting class selection (select period) that is still applicable.   

### Training and Testing Splits

Unfortunately, the ILEC data is aggregated in such a way that makes a typical train / test split cumbersome While we could sample the exposures and deaths for each aggregated cell, things like fractional exposures and the actuarial custom of setting deaths to have a full year of exposure introduce some opportunities for noise.  

Instead, we'll use the most recent two years of data for testing, and perform cross-validation within the training set by observation year.  This is more closely aligned with what an "unseen observation" is in practice, i.e., another calendar year of data.  


```{r}

df_extract_data <- df_study_data %>%
  mutate(
    dataset = case_when(
      (issue_year >= 2005) & (number_of_preferred_classes == "NA") ~ "simplified_issue",
      (issue_year >= 1998) & (number_of_preferred_classes != "NA") ~ "perm_recent",
      TRUE ~ "perm_historical"
    ),
    train_test = ifelse(observation_year >= 2016, "TEST", "TRAIN"),
    # no selection period, just ultimate mortality
    expected_deaths_08_vbt_ult = coalesce(qx * policies_exposed),
    # create ordered factor + integer representation of face amount bands
    face_amount_band = ordered(face_amount_band, levels=!!face_amount_order),
    face_amount_band_int = as.integer(face_amount_band)
  ) %>%
  filter(expected_deaths_08_vbt_ult > 0)

# create three extracts
df_extract_data %>%
  filter(dataset == "perm_historical") %>%
  write_rds("./ilec_perm_historical.rds", compress="gz")

df_extract_data %>%
  filter(dataset == "simplified_issue") %>%
  write_rds("./ilec_simplified_issue.rds", compress="gz")

df_extract_data %>%
  filter(dataset == "perm_recent") %>%
  write_rds("./ilec_perm_recent.rds", compress="gz")

```
