==> 3_mortality_improvement.Rmd <==
---
title: "Untitled"
author: "Mike McPhee Anderson, FSA"
date: "2025-01-01"
output: html_document
---

```{r setup, include=FALSE}

library(glmnet)
library(xgboost)
library(recipes)
library(rpart)
library(rpart.plot)
library(splines)
library(tidyverse)
library(GA)
library(memoise)

knitr::opts_chunk$set(echo = TRUE)

```


## Mortality Improvement

### Background

This [SOA Paper](https://www.soa.org/globalassets/assets/files/resources/essays-monographs/2017-living-to-100/2017-living-100-monograph-hunt-villegas-paper.pdf) provides a comprehensive overview of Poisson mortality improvement models.

The [Stochastic Mortality Modeling](https://cran.r-project.org/web/packages/StMoMo/vignettes/StMoMoVignette.pdf) package vignette provides a lot of additional background.

The approach taken here is very simple due to the relatively few observation years (10) contained in the data.  Consequently, we can only estimate the improvement over a 10 year period string in 2009.  The conceptual form of the model we'll use resembles the Poisson log-bilinear model:

$$ \ln \big[ \mu_{x,t} \big] = \pmb{X} \beta = \alpha_{x} + \gamma_t + \alpha_x \gamma_t $$

Where

 * $\alpha_x$ is the base mortality adjustment at age $x$
 
 * $\gamma_t$ is the all age mortality improvement rate at time $t$
 
 * $\alpha_x \gamma_t$ is the interaction effect (age-specific improvement) 

### Splines and knot selection

The above model form is overly simplistic if taken at face value.  Purely linear
terms, like $a_x$ are a poor fit for real mortality data. A *spline* is often used
to allow appropriate bends and wiggles in the shape of the line / curve for $a_x$.

Splines have parameters called *knots* that determine where additional "wiggliness"
is allowed in curves fit with GLMs. Adding knots increases the degrees of freedom
of the model, so having too many can lead to overfit or (worse) convergence issues
for the GLM fitting algorithm due to nearly-singular columns in the model matrix.

One approach to determining the right number of knots, and their location, is 
to use a *Genetic Algorithm*. Typically, knot placement are guessed then checked with
AIC / BIC or visual plots. Using a Genetic Algorithm speeds up the guessing and
checking process by getting us close to an optimal AIC/BIC automatically. This
approach is used below.

## Load Data

```{r}

ilec_perm <- bind_rows(
  read_rds("ilec_perm_historical.rds"),
  read_rds("ilec_perm_recent.rds"),
  read_rds("ilec_simplified_issue.rds")) %>% 
  filter(
    attained_age <= 100
  )

ilec_perm %>%
  select(attained_age, duration, observation_year) %>%
  summary()

```

### Create model data

```{r}

model_data <- ilec_perm %>%
  group_by(attained_age, duration, observation_year) %>%
  summarise(
    number_of_deaths = sum(number_of_deaths),
    policies_exposed = sum(policies_exposed),
    expected_deaths_08_vbt_ult = sum(expected_deaths_08_vbt_ult),
    .groups="drop"
  ) %>%
  ungroup() %>%
  mutate(
    observation_time = observation_year - min(observation_year) + 1
  )

```

### Setup the Genetic Algorithm Hypothesis Space

```{r}

source("0_lib.R")

duration_ptiles <- unique(approx_ecdf(model_data, duration, number_of_deaths)(seq(0.05, 0.95, 0.1)))
attage_ptiles <- unique(approx_ecdf(model_data, attained_age, number_of_deaths)(c(0.025, seq(0.1, 0.9, 0.1), 0.975)))
observation_time_ptiles <- unique(approx_ecdf(model_data, observation_time, number_of_deaths)(seq(0.25, 0.75, by=0.25)))

ATTAGE_PART = "splines::ns(attained_age, Boundary.knots = c(25, 100), knots=c(%KNOTS%))"
OBS_YR_PART = "splines::ns(observation_time, Boundary.knots = c(1, 10), knots=c(%KNOTS%))"
DUR_PART = "splines::ns(duration, Boundary.knots = c(1, 38), knots=c(%KNOTS%))"

formula_template <- stringr::str_glue("number_of_deaths ~ offset(log(expected_deaths_08_vbt_ult)) + {ATTAGE_PART}*{DUR_PART} + {ATTAGE_PART}*{OBS_YR_PART}")

spline_locations <- list()

spline_locations[[DUR_PART]] <- duration_ptiles
spline_locations[[ATTAGE_PART]] <- attage_ptiles
spline_locations[[OBS_YR_PART]] <- observation_time_ptiles 

formula_template

```

### Run the genetic algorithm

```{r}

N_BITS <- length(duration_ptiles) + length(attage_ptiles) + length(observation_year_ptiles)

bit_str_to_formula <- make_bit_str_to_formula(formula_template, spline_locations)

glm_fitness <- function(bit_string) {
  -1 * BIC(
    glm(bit_str_to_formula(bit_string), family = poisson(), data = model_data)
  )
}

# https://cran.r-project.org/web/packages/GA/vignettes/GA.html#memoization
best_model <- GA::ga(type="binary", 
       memoise::memoise(glm_fitness), 
       popSize=20, 
       nBits = N_BITS, 
       maxiter = 30, 
       parallel = F,
       pmutation = 0.7,
       elitism = 5,
       suggestions = rbind(
         rep(0, N_BITS),
         rep(1, N_BITS)
       ),
       pcrossover=0.7)

bit_str_to_formula(best_model@solution)

```

### Build model based on GA best solution

```{r}

# 128,345.0

final_formula <- number_of_deaths ~ offset(log(expected_deaths_08_vbt_ult)) + 
    splines::ns(attained_age, Boundary.knots = c(25, 100), knots = c(40, 50, 55, 60, 65, 70, 90)) * 
      splines::ns(duration, Boundary.knots = c(1, 38), knots = c(10, 17)) + 
    splines::ns(attained_age, Boundary.knots = c(25, 100), knots = c(40, 50, 55, 60, 65, 70, 90)) * observation_time

final_fit <- glm(final_formula, 
                 family = poisson(), 
                 data = model_data)

model_data[["MODEL_PRED"]] <- predict(final_fit, newdata=model_data, type = "response")

list(
  "AIC" = AIC(final_fit),
  "BIC" = BIC(final_fit),
  "D^2" = (final_fit.summary$null.deviance - final_fit.summary$deviance) / final_fit.summary$null.deviance,
  "Dispersion" = final_fit.summary$deviance / final_fit.summary$df.residual
)


```



### Check the fit

```{r fig.width=10}

attage_cuts <- c(0, seq(50, 80, by=10), Inf)

plot_data <- model_data %>%
  mutate(
    attage_grp = cut(attained_age, breaks=!!attage_cuts, ordered_result = T),
    duration = pmax(10, pmin(duration, 30))) %>%
  group_by(attage_grp, duration) %>%
  summarise(
    number_of_deaths = sum(number_of_deaths),
    MODEL_PRED = sum(MODEL_PRED),
    .groups="drop"
  ) %>%
  ungroup()


ggplot(plot_data, aes(x=duration, y=log(number_of_deaths / MODEL_PRED))) +
  facet_wrap(~ attage_grp, scales="free_y") +
  ylab("log A/E") +
  geom_point() +
  geom_line() + 
  theme_bw() +
  geom_hline(yintercept = 0, color="red")


```

```{r}

plot_data <- model_data %>%
  mutate(
    attage_grp = cut(attained_age, breaks=!!attage_cuts, ordered_result = T)) %>%
  group_by(attage_grp, observation_time) %>%
  summarise(
    number_of_deaths = sum(number_of_deaths),
    MODEL_PRED = sum(MODEL_PRED),
    .groups="drop"
  ) %>%
  ungroup()


ggplot(plot_data, aes(x=observation_time, y=log(number_of_deaths / MODEL_PRED))) +
  facet_wrap(~ attage_grp, scales="free_y") +
  ylab("log A/E") +
  geom_point() +
  geom_line() + 
  theme_bw() +
  geom_hline(yintercept = 0, color="red")


```

### Convert model to a set of multiplicative factors

```{r}

coef_groups <- tibble(
  coef_name = names(coef(final_fit))
) %>%
  mutate(coef_group = case_when(
    coef_name == "(Intercept)" ~ "Base Adjustment",
    stringr::str_detect(coef_name, fixed("attained_age")) & 
      stringr::str_detect(coef_name, fixed("duration")) ~ "Attained Age x Duration",
    stringr::str_detect(coef_name, fixed("attained_age")) & 
      stringr::str_detect(coef_name, fixed("observation_time")) ~ "Attained Age x Observation Time",
    stringr::str_detect(coef_name, fixed("attained_age")) ~ "Attained Age",
    stringr::str_detect(coef_name, fixed("observation_time")) ~ "Observation Time",
    stringr::str_detect(coef_name, fixed("duration")) ~ "Duration"
  )) 

any(is.na(coef_groups$coef_group))

```

### Mortality Improvement Factors

```{r}

factor_grid <- expand.grid(
  "attained_age" = seq(min(model_data$attained_age), max(model_data$attained_age)),
  "duration" = median(model_data$duration),
  "observation_time" = c(1, 2),
  "expected_deaths_08_vbt_ult" = c(1)
) %>%
  mutate(row_id = row_number())

factor_preds <- predict(final_fit, newdata=factor_grid, type = "terms")
colnames(factor_preds) <- paste0("factor_", colnames(factor_preds))

factor_grid.full <- bind_cols(
  factor_grid,
  factor_preds
) 

mi_factors <- factor_grid.full %>%
  rename(factor_attage_x_obs_time = `factor_splines::ns(attained_age, Boundary.knots = c(25, 100), knots = c(40, 50, 55, 60, 65, 70, 90)):observation_time`) %>%
  mutate(
    factor_full_mi = exp(factor_attage_x_obs_time + factor_observation_time)) %>%
  select(observation_time, attained_age, factor_full_mi) %>%
  pivot_wider(
    id_cols=c(attained_age),
    names_from = observation_time,
    values_from = factor_full_mi
  ) %>%
  mutate(
    mi_factor = `2` / `1`
  )

ggplot(mi_factors, aes(x=attained_age, y=mi_factor)) +
  geom_point() +
  geom_line() +
  theme_bw() +
  scale_y_continuous(labels = scales::percent_format(),
                     trans = scales::pseudo_log_trans(sigma=0.0001)) +
  scale_x_continuous(breaks = seq(25, 95, by=10)) +
  
  ylab("Improvement") +
  xlab("Attained Age") +
  ggtitle("Annual Mortality Improvement", "Ages 45-85 showing improvement")

```
