==> 7_mortality_gam_perm_recent.Rmd <==
---
title: "Perm Recent (GAM)"
author: "Mike McPhee Anderson, FSA"
date: "2025-05-10"
output: html_document
---

```{r setup, include=FALSE}


library(recipes)
library(rpart)
library(rpart.plot)
library(rsample)
library(scam)
library(mgcv)
library(plotly)
library(reticulate)
library(tidyselect)
library(tidyverse)

source("0_lib.R")
source("0_ga_lib.R")

#reticulate::use_miniconda("r-reticulate")

#optuna <- reticulate::import("optuna")

knitr::opts_chunk$set(echo = TRUE)

```


# Setup 

## Load Data

```{r}

face_amount_order <- c(
  "1-9999",
  "10000-24999",
  "25000-49999",
  "50000-99999",
  "100000-249999",
  "250000-499999",
  "500000-999999",
  "1000000-2499999",
  "2500000-4999999",
  "5000000-9999999",
  "10000000+"
)

df_study_data <- read_rds("ilec_perm_recent.rds") %>%
  mutate(
    train_test = ifelse(observation_year >= 2017, "TEST", "TRAIN"),
      # no selection period, just ultimate mortality
    expected_deaths_08_vbt_ult = coalesce(qx * policies_exposed),
    # create ordered factor + integer representation of face amount bands
    face_amount_band = ordered(face_amount_band, levels=!!face_amount_order),
    face_amount_band_int = as.integer(face_amount_band),
    issue_age = attained_age - duration + 1
  ) %>%
  filter(policies_exposed > 0)

```


# Perm Historical Model

## Train Test Split

```{r}

df_ph_data <- df_study_data 

df_ph_data.train <- df_ph_data %>% 
  filter(train_test == "TRAIN")

df_ph_data.test <- df_ph_data %>%
  filter(train_test == "TEST")

df_ph_data %>%
  group_by(train_test) %>%
  summarise(
    n_deaths = sum(number_of_deaths),
    .groups="drop"
  ) %>%
  ungroup() %>%
  mutate(
    pct_deaths = n_deaths / sum(n_deaths)
  )

```

## Data Prep

```{r}

group_model_data <- function(df) {
  df %>%
    group_by(attained_age, duration, issue_age, number_of_preferred_classes,
             preferred_class, smoker_status, face_amount_band_int, gender) %>%
    summarise(
      number_of_deaths = sum(number_of_deaths),
      policies_exposed = sum(policies_exposed),
      expected_deaths_08_vbt_ult = sum(expected_deaths_08_vbt_ult),
      .groups="drop"
    ) %>%
    ungroup()
}

step_my_model_factor_prep <- function(recipe) {
  recipe %>% step_mutate(
    issue_age = pmin(issue_age, 75),
    attained_age = pmin(attained_age, 95),
    # group 3-class with the 2-class, there isn't much of it
    preferred_class = case_when(
      # group super-pref & pref (3-class) as pref in the 2-class
      (number_of_preferred_classes == "3") & (preferred_class %in% c("1", "2")) ~ "1",
      # group standard as standard in the 2-class
      (number_of_preferred_classes == "3") & (preferred_class %in% c("3")) ~ "2",
      # set the NA (smoker distinct) classes to be "standard" in the two class, the
      # number of classes will be available to the model, but this gives it the option
      # of grouping standard in 2 class w/ the non-preferred classes, it also makes
      # default levels cleaner
      (number_of_preferred_classes == "NA") ~ "2",
      TRUE ~ preferred_class
    ),
    # set 3-class as 2-class, since applied the grouping above
    number_of_preferred_classes = case_when(
      number_of_preferred_classes == "3" ~ "2",
      number_of_preferred_classes == "NA" ~ "1",
      TRUE ~ number_of_preferred_classes
    ),
    # set factors w/ default levels
    #number_of_preferred_classes = factor(number_of_preferred_classes, levels=c("2")),
    preferred_class = factor(preferred_class, levels=c("2", "1")),
    smoker_status = factor(smoker_status, levels=c("NONSMOKER", "SMOKER")),
    gender = factor(gender, levels=c("MALE", "FEMALE"))
  ) 
}

model_data.train <- df_ph_data.train %>% group_model_data()
model_data.test <- df_ph_data.test %>% group_model_data()

ph.glmnet.recipe <- recipe(number_of_deaths ~ face_amount_band_int + 
                             attained_age + 
                             duration + 
                             issue_age +
                             gender +
                             number_of_preferred_classes +
                             preferred_class +
                             smoker_status + 
                             policies_exposed,
                           data = model_data.train) %>%
  step_my_model_factor_prep() %>%
  step_mutate(
    selection_group = interaction(smoker_status, preferred_class)
  ) %>%
  prep(model_data.train, retain=F)

model_data.train <- bake(ph.glmnet.recipe, model_data.train) 
model_data.test <- bake(ph.glmnet.recipe, model_data.test)

model_data.train %>%
  summary()

```

## Train GAM

```{r}

best_model_info <- environment()
best_model_info$best_score <- -Inf
best_model_info$best_model <- NULL

poisson_deviance <- carrier::crate(
  function(y, mu) {
    eps <- .Machine$double.eps
    y_safe <- pmax(y, eps)
    mu_safe <- pmax(mu, eps)
    terms <- y * (log(y_safe) - log(mu_safe)) - (y - mu)
    2 * sum(terms)
  })

d2_metric <- carrier::crate(
  function(y, mu_hat, mu_null) {
  
    if (length(mu_null) == 1) {
      mu_null <- rep(mu_null, length(y))
    }
    
    resid_dev <- poisson_deviance(y, mu_hat)
    null_dev  <- poisson_deviance(y, mu_null)
    
    100.0 * (null_dev - resid_dev) / null_dev
  },
  poisson_deviance=poisson_deviance
)

objective_fn <- carrier::crate(
  function(trial) {
    
    k_attage_gender <- trial$suggest_int("k_attage_gender", 4, 32, step=4)
    k_attage_sel_group <- trial$suggest_int("k_attage_sel_group", 4, 32, step=4)
    k_attage_issage_dur <- trial$suggest_int("k_attage_issage_dur", 4, 32, step=4)
    iter_gamma <- trial$suggest_float("gamma", -20, 0.5, step=0.5)
    
    iter_formula <- number_of_deaths ~ 
      stats::offset(log(policies_exposed)) + 
      preferred_class + 
      smoker_status + 
      s(attained_age, by=gender, k=k_attage_gender) +
      s(attained_age, by=selection_group, k=k_attage_sel_group) +
      ti(issue_age, duration, by=selection_group, k=k_attage_issage_dur) 
    
    fitting_error <- F
    tryCatch({
      fit <- mgcv::bam(
        iter_formula, 
        family = stats::poisson(), 
        data = model_data.train, 
        nthreads = 8,
        gamma = 2^iter_gamma,
        discrete=T)
    },
    error=function(e) {
      fitting_error <<- T  
    })
    
    perf_metric_val <- -Inf
    if (!fitting_error) {
      test_pred <- stats::predict(
        fit, newdata = model_data.test, type = "response"
      )
      
      perf_metric_val <- perf_metric(
        model_data.test$number_of_deaths,
        test_pred,
        sum(model_data.test$number_of_deaths) / sum(model_data.test$policies_exposed) * model_data.test$policies_exposed
      )  
    }
    
    if(best_model_info$best_score < perf_metric_val) {
      best_model_info$best_model <- fit
      best_model_info$best_score <- perf_metric_val
    }
    
    perf_metric_val
  },
  model_data.train = model_data.train,
  model_data.test = model_data.test,
  best_model_info = best_model_info,
  perf_metric = d2_metric,
  optuna = optuna)


```

```{r}

initial_params <- list(
  list(
    "gamma" = 0,
    "k_attage_gender" = 4,
    "k_attage_sel_group" = 4,
    "k_attage_issage_dur" = 4
  ),
  list(
    "gamma" = -6,
    "k_attage_gender" = 4,
    "k_attage_sel_group" = 4,
    "k_attage_issage_dur" = 4
  ),
  list(
    "gamma" = 0,
    "k_attage_gender" = 8,
    "k_attage_sel_group" = 8,
    "k_attage_issage_dur" = 8
  ),
  list(
    "gamma" = -6,
    "k_attage_gender" = 8,
    "k_attage_sel_group" = 8,
    "k_attage_issage_dur" = 8
  )
)

n_initial_params = length(initial_params)

#pruner <- optuna$pruners$MedianPruner(n_warmup_steps=5)

# Create a study with TPE sampler
study <- optuna$create_study(
  sampler = optuna$samplers$TPESampler(n_startup_trials=n_initial_params),
  #pruner = pruner,
  direction="maximize")

for (i in 1:n_initial_params) {
  study$enqueue_trial(initial_params[[i]])
}

#study$optimize(objective_fn, n_trials = 100L)

# Retrieve the best parameters
#best_params <- study$best_params
#print(best_params)

```

## Create final fit

```{r}

best_formula <- number_of_deaths ~ 
      stats::offset(log(policies_exposed)) + 
      preferred_class + 
      smoker_status + 
      s(attained_age, by=gender, k=8) +
      s(attained_age, by=selection_group, k=12) +
      ti(issue_age, duration, by=selection_group, k=12) 

fit <- mgcv::bam(
        best_formula, 
        family = stats::poisson(), 
        data = model_data.train, 
        nthreads = 8,
        gamma = 1.5,
        discrete=T)    

```


## Model Diagnostics

```{r fig.width=10}

model_data.train[["MODEL_PRED"]] <- predict(
  best_model_info$fit,
  newdata = model_data.train,
  type = "response"
)

plot_data <- model_data.train %>%
  group_by(duration, selection_group) %>%
  summarise(
    number_of_deaths = sum(number_of_deaths),
    MODEL_PRED = sum(MODEL_PRED)
  ) %>%
  pivot_longer(cols=-c(duration, selection_group))

ggplot(plot_data, aes(x=duration, y=value, color=name)) +
  facet_wrap(~ selection_group) +
  geom_line()

```

```{r fig.width=10}

model_data.test[["MODEL_PRED"]] <- predict(
  best_model_info$fit,
  newdata = model_data.test,
  type = "response"
)

plot_data <- model_data.test %>%
  group_by(duration, selection_group) %>%
  summarise(
    number_of_deaths = sum(number_of_deaths),
    MODEL_PRED = sum(MODEL_PRED)
  ) %>%
  pivot_longer(cols=-c(duration, selection_group))

ggplot(plot_data, aes(x=duration, y=value, color=name)) +
  facet_wrap(~ selection_group) +
  geom_line()

```

## Build full pred grid

```{r}

df_pred_grid <- with(model_data.train, {
  expand.grid(
    attained_age = seq(min(attained_age), max(attained_age)),
    duration = seq(min(duration), max(duration)),
    preferred_class = unique(preferred_class),
    smoker_status = unique(smoker_status),
    face_amount_band_int = seq(min(face_amount_band_int), max(face_amount_band_int)),
    gender = unique(gender)
  )
  }) %>%
  as_tibble() %>%
  mutate(
    issue_age = attained_age - duration + 1,
    policies_exposed = 1,
    selection_group = interaction(smoker_status, preferred_class)
  )

nrow(df_pred_grid)

```

```{r}

model_factor_data <- bind_cols(
  df_pred_grid,
  predict(
    fit,
    newdata=df_pred_grid,
    type = "terms") %>%
    as_tibble()) 

```

```{r}

plot_data <- model_factor_data %>%
  filter(
    issue_age >= 35,
    issue_age <= 75,
    gender == "MALE",
    selection_group == "NONSMOKER.1",
    face_amount_band_int == 5) %>%
  select(issue_age, duration, sel_factor = `ti(issue_age,duration):selection_groupNONSMOKER.1`)

grid_data <- plot_data %>%
  complete(duration = unique(plot_data$duration),
           issue_age = unique(plot_data$issue_age)) %>%
  arrange(issue_age, duration)

# Step 3: Pivot to wide format for z matrix
z_wide <- grid_data %>%
  pivot_wider(names_from = duration, values_from = sel_factor)

# Step 4: Extract axes and z matrix
z_matrix <- as.matrix(z_wide[,-1])  # remove issue_age column
y_vals <- z_wide$issue_age          # row axis: issue_age
x_vals <- sort(unique(plot_data$duration))  # column axis: duration

# Step 5: Plot surface
fig <- plot_ly(
  x = x_vals,
  y = y_vals,
  z = z_matrix,
  type = "surface",
  colorscale = "Viridis",
  showscale = TRUE
) %>%
  layout(
    scene = list(
      xaxis = list(title = "Duration"),
      yaxis = list(title = "Issue Age"),
      zaxis = list(title = "Select Factor")
    ),
    title = "3D Select Factor Surface"
  )

fig
```

### Fit constrained model

#### Prepare data

```{r}

mutate_sel_grp <- function(df) {
  df %>%
    mutate(
      issage_grp = cut(
        issue_age,
        breaks = c(-Inf, seq(35, 65, by=10), Inf)
      ),
      selection_group_w_issage = interaction(
        selection_group, issage_grp
      )
    )
}

model_data.train <- model_data.train %>%
  mutate_sel_grp()

model_data.test <- model_data.test %>%
  mutate_sel_grp()

```

```{r}

model_data.train %>%
  group_by(selection_group_w_issage, duration) %>%
  summarise(
    number_of_deaths = sum(number_of_deaths),
    .groups="drop"
  ) %>%
  ungroup() %>%
  pivot_wider(
    id_cols = duration, 
    names_from = selection_group_w_issage,
    values_from = number_of_deaths,
    values_fill = 0) %>%
  view()

```


```{r}

constrained_model_formula <- number_of_deaths ~ 
      stats::offset(log(policies_exposed)) + 
      preferred_class + 
      smoker_status + 
      s(attained_age, by=gender, k=8) +
      s(attained_age, by=selection_group, k=12) +
      s(pmin(15, duration), by=selection_group_w_issage, k=8, bs="mifo") 

```

```{r}

constrained_fit <- scam(
        constrained_model_formula, 
        family = stats::poisson(), 
        data = model_data.train, 
        gamma = 0.5)    

summary(constrained_fit)

```

```{r}

df_pred_grid <- df_pred_grid %>% 
  mutate_sel_grp()

df_term_preds <-  predict(
  constrained_fit,
  newdata=df_pred_grid,
  type = "terms") %>%
  as_tibble() %>%
  mutate(
    row_id = row_number()
  ) %>%
  pivot_longer(
    cols = -c(row_id),
    names_to = "term_name",
    values_to = "term_value"
  ) %>%
  mutate(
    term_group = case_when(
      stringr::str_detect(term_name, "^s\\(attained_age\\).*") ~ "factor_attage",
      stringr::str_detect(term_name, "^s\\(pmin\\(15, duration\\)\\).*") ~ "factor_duration",
      TRUE ~ "other"
    )
  ) %>%
  group_by(row_id, term_group) %>%
  summarise(
    term_value = exp(sum(term_value)),
    .groups="drop"
  ) %>%
  pivot_wider(id_cols = row_id, 
              names_from = "term_group", 
              values_from = "term_value",
              values_fill = 0) %>%
  arrange(row_id)

df_term_preds
```

```{r}

constrained_model_factor_data <- bind_cols(
  df_pred_grid,
  df_term_preds
 ) 

```


```{r}
colnames(constrained_model_factor_data)
```

# Manually Adjust Factors

```{r}

constrained_model_factor_data %>%
  select(selection_group_w_issage, duration, factor_duration) %>%
  distinct() %>%
  pivot_wider(id_cols = duration, 
              names_from = selection_group_w_issage,
              values_from =  factor_duration) %>%
  write_csv("sel_factors.csv")

```


```{r}

plot_data <- model_factor_data %>%
  filter(
    issue_age >= 35,
    issue_age <= 75,
    gender == "MALE",
    selection_group == "NONSMOKER.1",
    face_amount_band_int == 5) %>%
  select(issue_age, duration, sel_factor = `ti(issue_age,duration):selection_groupNONSMOKER.1`)

```
